A method is extracted, and a change is performed.
A method is extracted to remove duplicate code, and a change is performed on the detectRefactoring method to consider the commits contains only added or removed. Although there is no refactoring in such a commit, we need them in case we want to find added or removed code elements like classes or methods.